+++ src/impls/smallvec.rs
@@ -1,5 +1,5 @@
 use crate::{
-	Back, BackMut, Capacity, Clear, Collection, CollectionMut, CollectionRef, Front, FrontMut, Len,
+	Capacity, Clear, Collection, CollectionMut, CollectionRef, Get, GetMut, Iter, IterMut, Len,
 	PopBack, PushBack, Remove, Reserve, WithCapacity,
 };
 use smallvec::{Array, SmallVec};
@@ -42,59 +42,51 @@ impl<A: Array> Len for SmallVec<A> {
 }
 
 impl<A: Array> Capacity for SmallVec<A> {
+	#[inline(always)]
 	fn capacity(&self) -> usize {
 		self.capacity()
 	}
 }
 
 impl<A: Array> Reserve for SmallVec<A> {
+	#[inline(always)]
 	fn reserve(&mut self, additional: usize) {
 		self.reserve(additional)
 	}
 }
 
-impl<A: Array> Back for SmallVec<A> {
-	fn back(&self) -> Option<&A::Item> {
-		self.last()
+impl<A: Array> Get<usize> for SmallVec<A> {
+	#[inline(always)]
+	fn get(&self, index: usize) -> Option<&A::Item> {
+		self.as_slice().get(index)
 	}
 }
 
-impl<A: Array> BackMut for SmallVec<A> {
-	fn back_mut(&mut self) -> Option<&mut A::Item> {
-		self.last_mut()
-	}
-}
-
-impl<A: Array> Front for SmallVec<A> {
-	fn front(&self) -> Option<&A::Item> {
-		self.first()
-	}
-}
-
-impl<A: Array> FrontMut for SmallVec<A> {
-	fn front_mut(&mut self) -> Option<&mut A::Item> {
-		self.first_mut()
+impl<A: Array> GetMut<usize> for SmallVec<A> {
+	#[inline(always)]
+	fn get_mut(&mut self, index: usize) -> Option<&mut A::Item> {
+		self.as_mut_slice().get_mut(index)
 	}
 }
 
 impl<A: Array> PushBack for SmallVec<A> {
-	type Output<'a>
-	where
-		Self: 'a,
-	= ();
+	type Output = ();
 
+	#[inline(always)]
 	fn push_back(&mut self, t: A::Item) {
 		self.push(t)
 	}
 }
 
 impl<A: Array> PopBack for SmallVec<A> {
+	#[inline(always)]
 	fn pop_back(&mut self) -> Option<A::Item> {
 		self.pop()
 	}
 }
 
 impl<A: Array> Remove<usize> for SmallVec<A> {
+	#[inline(always)]
 	fn remove(&mut self, index: usize) -> Option<A::Item> {
 		if index < self.len() {
 			Some(self.remove(index))
@@ -105,7 +97,32 @@ impl<A: Array> Remove<usize> for SmallVec<A> {
 }
 
 impl<A: Array> Clear for SmallVec<A> {
+	#[inline(always)]
 	fn clear(&mut self) {
 		self.clear()
+	}
+}
+
+impl<A: Array> Iter for SmallVec<A> {
+	type Iter<'a>
+	where
+		A: 'a,
+	= std::slice::Iter<'a, A::Item>;
+
+	#[inline(always)]
+	fn iter(&self) -> Self::Iter<'_> {
+		self.as_slice().iter()
+	}
+}
+
+impl<A: Array> IterMut for SmallVec<A> {
+	type IterMut<'a>
+	where
+		A: 'a,
+	= std::slice::IterMut<'a, A::Item>;
+
+	#[inline(always)]
+	fn iter_mut(&mut self) -> Self::IterMut<'_> {
+		self.as_mut_slice().iter_mut()
 	}
 }