+++ CHANGELOG.md
@@ -9,6 +9,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added 
 - Required `Clone` traits on immutable reference types.
 
-## [0.5.0] - 2021-09-22
+## [0.5.2] - 2021-09-26
+### Added
+- `Get` and `GetMut` impls for `SmallVec`.
+
+## [0.5.1] - 2021-09-26
+### Changed
+- Fix features `slab` and `smallvec`.
+
+## [0.5.0] - 2021-09-22 [YANKED]
+### Yanking reason
+- Errors with features `slab` and `smallvec`.
+
 ### Added 
 - `Keyed` and `KeyedRef` traits.